HDDS-8817. Create separate check for native build (#4875)
