+++ .github/workflows/ci.yml
@@ -71,8 +71,6 @@ jobs:
     steps:
       - name: Checkout project
         uses: actions/checkout@v3
-        with:
-          fetch-depth: 0
       - name: Cache for npm dependencies
         uses: actions/cache@v3
         with:
@@ -153,7 +151,6 @@ jobs:
         run: hadoop-ozone/dev-support/checks/build.sh -Dskip.npx -Dskip.installnpx -Djavac.version=${{ matrix.java }}
         env:
           OZONE_WITH_COVERAGE: false
-          SKIP_NATIVE_VERSION_CHECK: true
       - name: Delete temporary build artifacts before caching
         run: |
           #Never cache local artifacts
@@ -173,12 +170,12 @@ jobs:
     steps:
       - name: Checkout project
         uses: actions/checkout@v3
-        if: matrix.check != 'bats' && matrix.check != 'unit'
+        if: matrix.check != 'bats'
       - name: Checkout project with history
         uses: actions/checkout@v3
         with:
           fetch-depth: 0
-        if: matrix.check == 'bats' || matrix.check == 'unit'
+        if: matrix.check == 'bats'
       - name: Cache for maven dependencies
         uses: actions/cache@v3
         with:
@@ -338,10 +335,8 @@ jobs:
           - flaky
       fail-fast: false
     steps:
-      - name: Checkout project with history
+      - name: Checkout project
         uses: actions/checkout@v3
-        with:
-            fetch-depth: 0
       - name: Cache for maven dependencies
         uses: actions/cache@v3
         with:
