+++ dev-support/ci/selective_ci_checks.sh
@@ -82,7 +82,7 @@ function get_changed_files() {
 }
 
 function set_outputs_run_everything_and_exit() {
-    BASIC_CHECKS="author bats checkstyle docs findbugs rat unit"
+    BASIC_CHECKS="author bats checkstyle docs findbugs native rat unit"
     compile_needed=true
     compose_tests_needed=true
     dependency_check_needed=true
@@ -434,6 +434,23 @@ function check_needs_findbugs() {
     start_end::group_end
 }
 
+function check_needs_native() {
+    start_end::group_start "Check if native is needed"
+    local pattern_array=(
+        "^hadoop-ozone/dev-support/checks/native.sh"
+        "^hadoop-hdds/rocks-native"
+        # include tests tagged as @Native in any module
+        $(grep -Flr 'org.apache.ozone.test.tag.Native' hadoop-*/*/src/test/java)
+    )
+    filter_changed_files true
+
+    if [[ ${match_count} != "0" ]]; then
+        add_basic_check native
+    fi
+
+    start_end::group_end
+}
+
 function check_needs_unit_test() {
     start_end::group_start "Check if unit test is needed"
     local pattern_array=(
@@ -591,6 +608,7 @@ check_needs_checkstyle
 check_needs_dependency
 check_needs_docs
 check_needs_findbugs
+check_needs_native
 check_needs_unit_test
 calculate_test_types_to_run
 set_outputs
