+++ hadoop-hdds/rocks-native/pom.xml
@@ -27,6 +27,10 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.ozone</groupId>
+            <artifactId>hdds-test-utils</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.ozone</groupId>
             <artifactId>hdds-managed-rocksdb</artifactId>
         </dependency>
         <dependency>
@@ -460,25 +464,6 @@
                                 </configuration>
                             </execution>
                         </executions>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>disable-native-testing</id>
-            <activation>
-                <property>
-                    <name>!rocks_tools_native</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <groupId>org.apache.maven.plugins</groupId>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <configuration>
-                            <excludedGroups>native</excludedGroups>
-                        </configuration>
                     </plugin>
                 </plugins>
             </build>
