+++ hadoop-ozone/dev-support/checks/build.sh
@@ -26,12 +26,6 @@ else
   MAVEN_OPTIONS="${MAVEN_OPTIONS} -Djacoco.skip"
 fi
 
-if [[ "${SKIP_NATIVE_VERSION_CHECK}" != "true" ]]; then
-  NATIVE_MAVEN_OPTIONS="-Drocks_tools_native"
-  . "$DIR/native_check.sh"
-  init_native_maven_opts
-  MAVEN_OPTIONS="${MAVEN_OPTIONS} ${NATIVE_MAVEN_OPTIONS}"
-fi
 export MAVEN_OPTS="-Xmx4096m $MAVEN_OPTS"
 echo "${MAVEN_OPTIONS}"
 mvn ${MAVEN_OPTIONS} clean install "$@"
