+++ hadoop-ozone/fault-injection-test/network-tests/pom.xml
@@ -27,6 +27,10 @@ http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <name>Apache Ozone Network Tests</name>
   <packaging>jar</packaging>
 
+  <properties>
+    <maven.test.skip>true</maven.test.skip> <!-- no tests in this module so far -->
+  </properties>
+
   <build>
     <plugins>
       <plugin>
