+++ hadoop-ozone/httpfsgateway/pom.xml
@@ -36,6 +36,7 @@
     <httpfs.source.revision>REVISION NOT AVAIL</httpfs.source.revision>
     <maven.build.timestamp.format>yyyy-MM-dd'T'HH:mm:ssZ</maven.build.timestamp.format>
     <httpfs.build.timestamp>${maven.build.timestamp}</httpfs.build.timestamp>
+    <maven.test.skip>true</maven.test.skip> <!-- no tests in this module so far -->
   </properties>
 
   <dependencies>
