+++ hadoop-ozone/ozonefs-hadoop3-client/pom.xml
@@ -38,6 +38,9 @@
       <optional>true</optional>
     </dependency>
   </dependencies>
+  <properties>
+    <maven.test.skip>true</maven.test.skip> <!-- no tests in this module so far -->
+  </properties>
   <build>
     <plugins>
       <plugin>
